# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Angular Workshop API",
        "description": "API documentation for Angular Workshop application with categories, skills, exams, and user management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get all skills",
                "operationId": "abc589672de509c5e1ae0eb9f43d7f65",
                "responses": {
                    "200": {
                        "description": "List of skills",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Skill"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Create a new skill",
                "operationId": "0d51b818cd83d1f12dcc8a56345c3dbe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_en",
                                    "name_ar",
                                    "cat_id"
                                ],
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "name_ar": {
                                        "type": "string",
                                        "example": "لارافيل"
                                    },
                                    "cat_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "img": {
                                        "type": "string",
                                        "example": "laravel.png"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skill"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/skills/{id}": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get a specific skill with exams",
                "operationId": "f9d72df8282e1223cebfcfe15ab06c65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skill"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Skills"
                ],
                "summary": "Update a skill",
                "operationId": "606b380c07b19fdbe3c65b395a5dbbb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name_en": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "name_ar": {
                                        "type": "string",
                                        "example": "لارافيل"
                                    },
                                    "cat_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "img": {
                                        "type": "string",
                                        "example": "laravel.png"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Skill"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Skill not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Skills"
                ],
                "summary": "Delete a skill",
                "operationId": "87b15955e71b8374d6e9509081c24620",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skill deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Skill": {
                "title": "Skill",
                "description": "Skill model",
                "required": [
                    "id",
                    "name",
                    "cat_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "object",
                        "example": {
                            "en": "Laravel",
                            "ar": "لارافيل"
                        }
                    },
                    "cat_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}